npm, gulp, and other build stuff
package.json file
gulp
setup gitignore.
browserify: figuring out standalone project.
moving code around into different module files.
add d3 js as a node module.
holy cow it actually worked.
