+++ index.html
@@ -10,14 +10,9 @@
 /* ]]> */
 </style>
 -->
-
-<script type="text/javascript" src="https://cdnjs.cloudflare.com/ajax/libs/underscore.js/1.7.0/underscore-min.js"></script>
-<script type="text/javascript" src="https://cdnjs.cloudflare.com/ajax/libs/d3/3.5.2/d3.min.js"></script>
-<script type="text/javascript" src="core.js"></script>
+<script type="text/javascript" src="dist/asset/js/genomic-bundle.js"></script>
 </head>
 <body>
-
     <div id="main"></div>
-
 </body>
 </html>
