+++ src/js/utils.js
@@ -1,0 +1,5 @@
+var translate = function(x,y) {
+  return "translate(" + x + "," + y + ")";
+}
+
+exports.translate = translate;
